refactor(plugins): extract Plugin interface to api package and update registry to use it
