+++ api/plugin.go
@@ -1,0 +1,8 @@
+package api
+
+import "github.com/urfave/cli/v3"
+
+type Plugin interface {
+	Name() string
+	Register(root *cli.Command)
+}
