+++ internal/plugins/registry_test.go
